+++ src/main/java/young/board/post/service/PostResponseDto.java
@@ -1,18 +1,16 @@
-package young.board.post;
+package young.board.post.service;
 
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import young.board.domain.Category;
-import young.board.domain.Comment;
 import young.board.domain.Post;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class PostResponse {
+public class PostResponseDto {
 
     private Long id;
     private Category category;
@@ -20,12 +18,12 @@ public class PostResponse {
     private String writer;
     private LocalDateTime createDate;
     private String content;
-    private List<Comment> comments;
+//    private List<Comment> comments; //음... 고민 해보자 다른 방법은 없을까?
 
-    private Integer likeNumber;
+//    private Integer likeNumber;
 
-    public static PostResponse create(Post post, Integer calculateLikesCnt, List<Comment> comments) {
-        PostResponse postResponse = new PostResponse();
+    public static PostResponseDto from(Post post) {
+        PostResponseDto postResponse = new PostResponseDto();
         postResponse.id = post.getId();
         postResponse.title = post.getTitle();
         postResponse.category = post.getCategory();
@@ -33,9 +31,9 @@ public class PostResponse {
         postResponse.writer = post.getWriter();
         postResponse.content = post.getContent();
         postResponse.createDate = post.getCreateDate();
-        postResponse.likeNumber = calculateLikesCnt;
-        postResponse.comments = comments;
-
+//        postResponse.likeNumber = calculateLikesCnt;
+//        postResponse.comments = comments;
         return postResponse;
     }
+
 }
